+++ analysis/import/SonarImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/sonar/SonarMeasuresAPIImporter.kt
@@ -23,12 +23,19 @@ class SonarMeasuresAPIImporter @JvmOverloads constructor(
         val componentMap = measuresDS!!.getComponentMap(projectKey, metricsList)
 
         val projectBuilder = SonarComponentProjectBuilder(sonarCodeURLLinker, translator, usePath)
+        val attributeDescriptors = getAttributeDescriptors()
+        projectBuilder.addAttributeDescriptions(
+            attributeDescriptors
+                .filter { it.key in metricsList }
+                .map { translator.translate(it.key) to it.value }
+                .toMap()
+        )
         return projectBuilder.addComponentMapsAsNodes(componentMap).build()
     }
 
     fun getMetricList(metrics: List<String>): List<String> {
-        return if (metrics.isEmpty()) {
+        return metrics.ifEmpty {
             metricsDS!!.availableMetricKeys
-        } else metrics
+        }
     }
 }